requirements.txt
streamlit
yfinance
pandas
numpy
plotly
requests
polygon-api-client
edgar
openpyxl